requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.0
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
drf-yasg==1.21.7
python-decouple==3.8
psycopg2-binary==2.9.9
